+++ src/apps/api/src/graphql/mutations/index.ts
@@ -1,5 +1,7 @@
 import * as users from "./user";
+import * as reviews from "./review"
 
 export const Mutation = {
-  ...users
+  ...users,
+  ...reviews
 };
