+++ src/apps/api/src/graphql/mutations/review/index.ts
@@ -1,1 +1,1 @@
-export * from "./create";
+export * from "./createReview";
