+++ src/apps/api/src/graphql/queries/index.ts
@@ -1,5 +1,6 @@
 import * as users from "./user";
+import * as reviews from "./review"
 
 export const Query: any = {
-  ...users,
+  ...users
 };
